PicChat: finish Authentication reading

+++ projects/pic_chat/lib/pic_chat_web/live/message_live/index.ex
@@ -45,8 +45,17 @@ defmodule PicChatWeb.MessageLive.Index do
   @impl true
   def handle_event("delete", %{"id" => id}, socket) do
     message = Messages.get_message!(id)
-    {:ok, _} = Messages.delete_message(message)
 
-    {:noreply, stream_delete(socket, :messages, message)}
+    if message.user_id == socket.assigns.current_user.id do
+      {:ok, _} = Messages.delete_message(message)
+      {:noreply, stream_delete(socket, :messages, message)}
+    else
+      {:noreply,
+       Phoenix.LiveView.put_flash(
+         socket,
+         :error,
+         "You are not authorized to delete this message."
+       )}
+    end
   end
 end
